Fix all tests
This added a new function `conditions.disable_globaltag_replay()` as
convenience in case one wants to restore the old default behavior. That
was doable without this function but more cumbersome.

For mdst tests: Two of the older releases change the values because the
magnetic field in their globaltag is different to what we use now. The
correct thing to do is use the original globaltag.

I also rewrote the `analysis/tests/tutorials.py` to only output the log
of the steering file if it fails, otherwise it is too noisy to find
errors.