B2ERROR, when nnt weights are not loaded correctly
when no filename is given to the neurotrigger, it tries to load the
network weights from the conditions database. this worked in the past,
but for some reason the payloads for the neurotrigger disappeared now.
To fix this, an error is thrown now if there are no network weight sets.