Add module and script to convert Benergy and IPProfile from Belle data
This adds a module to do the conversion of IPProfile and Benergy from
Belle format to Belle II conditions payloads. This does basically the
same as B2BIIConvertMdst module except that it does just create payloads
in the begin run method and doesn't need input files: It will fudge the
belle run records so that the correct values are obtained from the
database.

It also adds a command line tool to run the conversion. In the simplest
case one can just run

```
b2bii-convert-belle-ipprofile -j16
```

to run 16 parallel jobs doing the conversion. If the program is
interrupted one can just rerun it in the same folder with the same
arguments and it will continue where it left off (unless --delete is
specified).

Once running the script says nothing is to be done anymore the payloads
are in principle ready to be uploaded to the database. However there
will be multiple runs with the same contents for some of them so it
might be advisable to to some postprocessing on the database file to
combine iovs where possible.

The postprocessing script doesn't exist yet but can be done quickly once
we have the payloads.

This script needs access tho the BELLE postgres database so it can only
be run at KEKCC (unless proper port forwarding is setup)