Re-add accidentally removed module in default path: Non-MVA VXD QE
The conditional path at the end of `add_vxd_track_finding_vxdtf2` was very
confusing to read, so I rewrote it in a simpler and imo more expressive way. It
seems however that I forgot to add the module `QualityEstimatorVXD` if no MVA
quality estimator is used, as that module is still used for cutting. I found
this error only out when running the validation and where I saw a detoriation in
the SVD efficiencies.