Use more expressive names for temporary SVD-CDC-standalone tracks
In the function `add_svd_track_finding()` it is not immediately clear what
`temporary_reco_tracks_two` means and what it is for. Also, the assigned default
string `"SVDplusRecoTracks"` does not tell the reader of the code immediately
what is contained in the store array. In particular, it is not clear that this
store array contains CDC tracks found by the CDC standalone track finding. It is
basically just the old `"SVDCDCRecoTracks"` store array before the SVD-to-CDC
CKF was added.