Make sure conditions configuration reflects what is used
Currently it's possible to modify the conditions configuration object
after the database has been initialized and these changes will not be
reflected by the database.

This change adds a write protection to the configuration once the
database is initialized. Users can call `conditions.reset()` to reset
the database and the configuration object in this case.

In this case users cannot accidentally cange the configuration as they
have to explictly reset first. But if they want to change the
configuration after they database is initialized they can do so.

Effect on users should be small: Almost nobody has complicated enough
use cases of the database configuration to notice this change, only some
scripts for the softwaretrigger try to initialize the database twice
with different settings.