Change TestingPayloadStorage to use md5 as revision
* Change the reading part to not convert the revision to an integer but
  to just keep it as it is. Also, reverse the list of iovs per payload
  and always use the last one in the file, not the highest revision.

* Change creation of payloads to create payloads in a temporary file
  first and then rename them to their final location afterwards. Replace
  the revision with an abbreviated md5 sum of the payload.

This gets rid of race conditions when multiple processes would write to
different payload text files in the same folder and also optimizes the
time it needs to find a new payload in a folder with large amount of
payload files.

It also makes it more clear that these are local payload files and not
from a central database and avoids creating new payload files if the
same content already exists.