b2conditionsdb dump: Add way to dump payload valid for certain run
In addition to allow dumping files, revisions and payload ids one can
now also dump a payload with a given name, globaltag, run, exp and the
first payload. It will then dump the payload content of the payload
valid at this time.

Also, allow to dump payload files which have checksum as revision.