Add/refactor QE default module params and remove them at call sites
E.g. when always the "tripletFit" estimation method is used for the VXDTF2
quality estimator, add that as the default value for the module parameter and
remove that specification from the helper function / steering file where the
module is called.

Explicit is better than implicit, but only when it provides useful information
for the user and defaults still don't hurt.