Allow use of metavariables in create_for_selected (BII-4896)
This allows providing meta variables for create_for_selected et. al. by
just replacing all non alphanumeric characters in the variable name with
`_` in the alias name.

Somehow we missed this for release-04 but I think we should push it in
the release branch, we'll need a patch release anyhow.