Change QE teacher create both .xml and .root weightfiles
The trained xml weightfiles can be used by the MVA quality estimator classifiers
locally via the path to the weightfile and are therefore used internally by this
training script.

However, the weightfile a requires in the ROOT format, if one wants two upload
them to the conditions database. I solved this by just doing the training twice
for each classifier, once with a ".xml" and once with a ".root" file ending.
As fastbdt is really quick in training, this should not be an issue. However, it
would be nice if there were conversion methods...