Add default-enabled option to export QI additionally to mdst Tracks
Fitted mdst Tracks have already a float m_qualtiyIndicator property, but it is const
and never set. After compression, it probably didn't take up much space in the
ntuples, but this is changed when it will be set to a varying number. Therefore,
adding the QI to the Tracks in addition to the final RecoTracks will cost storage/money.