Add a script to calculate all beam related payloads
Since release 4 it's no longer possible to analyse MC which used a
different beam energy.

This is an oversight in release 4: We anyway want to move away from
forcing the user to specify beam energies in the steering file but this
part was just not considered.

With this script we can create the necessary payloads from the
beamparameter preset similar to before. These payloads could then either
be used locally using

    basf2.conditions.testing_payloads = ['localdb/database.txt']

but we can also upload them to dedicated globaltags and then just add
these globaltags in the MC production scripts for non-4S events and then
global tag replay will do the rest.

This script can estimate the uncertainties of the invariant mass and the
boost vector by doing a simple toy mc using the beam parameters. It
doesn't account for possible differences between generated/reconstructed
as we do for the Beamspot for 4S MC.

We could add a constant fudge factor to the beam spot size from what we
know from 4S to make it more conssitent but I'm not familiar enough with
that.