Try to make hlt zmq tests more reliable (BII-5829)
The current tests fail most of the time on bamboo. The reason is most
probably that there's a too long delay between determining a port number
and actually binding to the port to make it exclusive.

This commit fixes test_worker.py by opening the sockets directly and
just using them so there is no delay at all: Sockets are used and then
handed to the program.

For all other tests it moves the determination of the port number as
close to the start of the programs using them as possible, that is in
the setUp() instead on static class members.

This is still not 100% race condition free but should be much more
robust and these tests didn't exhibit the large failure rate observed by
the worker test.