Add experiment_numbers setting to iterate over
For training for different geometries and experimental setups. Be aware that
this setting does not changed that background, that should be changed
accordingly. E.g. if you choose experiment 1002, then use phase2 background.

Also the reconstruction chain is not changed based on the experiment number,
except where the helper functions do it automatically.