Fix empty globaltag in output metadata if database is not used (BII-5790)
When the database is not used the list of globaltags was never
initialized so the list of globaltags in the output metadata was empty.

We now make sure the list of globaltags is initialized early but we keep
the final initialization of the database communication still as late as
possible (or not at all).