change default neurotrigger unpacker parameters
correct the delay fields, now all information is written out in the same
cc
add new field for output scaling: the hardware output is just a
normalized value where the scale factor depends on the network training
settings. if this field is left empty, it defaults to the values of the
old background network Background2.0_20161207.root. It can also be set
manually. In case it is set to an empty list, the values are loaded from
the correct network stored in the conditions database.