Allow to filter repeated log messages
This commit adds a new setting to the logging system to suppress
identical log messages. Here identical means "same text and same level"
but excludes variables, package, module, line and file information.

The (current) default is to not suppress any messages but one can set

    basf2.logging.max_repetitions = 100

to only save the first 100 messages with the same text. All subsequent
messages with the same text are then omitted, independent of what other
messages are emitted (currently we only filter repeated messages if they
occur directly after each other).

The logging system will print a short message from time to time if
messages are suppressed (each 100 suppressed messages for the first
1000, then each 1000 for the first 10000 and then each 10000 suppressed
messages).

The log summary will contain the correct message count for each message.
If a message was shown more than once the log summary will not contain
the variables associated with the message as these might have changed
between each message. Also the log summary is now sorted by severty and
count and not chronologically (since we don't have that information
anymore).

In addition I made the message when processing is interrupted by SIGINT
more verbose and expanded the documentation for `set_log_level()`

Limitations:

 * There is no way to reenable logging of the suppressed messages except
for resetting the logging.
 * If there are more than 10k different messages suppression doesn't
   work for additional messages (to not allow growing to infinite size)
   so these messages will all be shown. Risk for this can be mitigated
   if we make sure variables are used consistently for log messages.
 * The log summary is nor formatted slightly differently (just to annoy
   people using regex on the log summary ;))