Inline methods to check if log level is enabled
I did some quick measurements and this can speed up the check if a log
message is enabled significantly. So especially if there are many debug
messages in a loop this can have a sizeable effect and it's just
movement of code to a position the compiler can always see. (LTO would
not necessarily help as this function would then just be inlined in
libframework.so but not across library boundaries)