VariableManager: Enable recursive alias lookup (BII-6118)
When resolving a variable we only looked for an alias mapping exactly
once. That means an alias to an alias didn't work.

That might be fine as a design decision but due to the lookup being done
in steps an alias to a metavariable to an alias does work. Which makes
it a very inconsistent user experience.

This commit changes the alias lookup to be performed recursively until
no more aliases are found but also makes sure that loops in the lookup
are detected. As the variable lookup is usually done exactly once at the
first use of a variable this should not affect performance.