Allow merging of files if Legacy_IP_Information is inconsistent among inputs (BII-6229)
We sometimes added the Legacy_IP_Information to the list of globaltags
even though it's not really needed (as it doesn't hurt). This causes
problems that we might not be able to merge files if this is
inconsistent.

This change checks if the inconsitency in globaltags is just caused by
Legacy_IP_Information and if so removes it: We know part of the files
were processed without it so the other globaltags contain all the
information we need.