Add a EParticleType::c_V0 and propagate this.
Resolves BII-6331 and BII-3698.
Particles constricted from V0s were previously completely detached from
the V0 object post construction, and were treated as composite
particles. This is not terrible, but it means they are in disinguishable
from Kshorts, Lambdas, etc created by the ParticleCombiner.
This has (positive) implications for the V0 standard lists and for the V0Merger.