Allow escaping of newlines in log messages (BII-6470)
The online system currently only retains the first line of each log message. Fixing this by collecting all lines until the next log level marker could lead to unintended side effects in other parts of the online system. So as a quick solution we allow to set escaping of newlines in the logmessage to \n. This will put the whole message on one line and the change on he online system to convert these escapes back without side effect is trivial.