Allow b2conditionsdb-tag-modify to change only the state (BII-6509)
When modifying a tag that was not OPEN the modification would fail as
the state cannot be modified. However this would fail even if no values
would actually be modified.

Now we only attempt to change the tag if there's an actual change of
values so a `b2conditionsdb tag modify -s OPEN foo` will always work
(assuming the state transition is allowed)