Add tools to check class versions and checksums
Add the two new tools `b2code-classversion-check` and
`b2code-classversion-update` which should help users to avoid errors
when updating classes to be saved in ROOT files.

The check tool can

* check that a class is fully defined (no forward declarations)
* check that all its members have a dictionary
* compare class version and checksum with the values in the comment
after the class link pragma

The version/checksum values are put into the comment as comma separated
list and will also work fine with the `implicit` keyword

The update tool can update the linkdef files to automatically put the
correct values in the linkdef file after the user fixed the ClassDef
version number.