Fix Problem with event generation introduced by PR #5960 (BII-6783)
The idea was to use the beam parameters to smear the vertex but my
request to handle already smeared vertices correctly was a bit rushed.
So this pr fixes the regression observed in evtgen.

1. Revert the changes to the initial particle generator
2. Add a new function to update just the vertex.
   * if there was no previous initial particles create some but without
     smearing of the beams as the generator apparently didn't use the smearing
   * if vertex smearing is allowed and not already applied, smear the vertex
   * return the shift between the previous value and the new vertex. If
     there was no previous value then the shift wrt to the origin.

This sounds like a new feature but what was planned for PR #5960 but not
implemented correctly. The original function is reverted to it's old
form to fix the regression which might have also affected other
generators. This new function is only used by the smear primary vertex
module.