Propagate particles to the top volume before giving them to Geant4
Some generators (for example cosmics) might generate particles outside
of the simulation volume. This currently causes a crash/abort as Geant4
doesn't like to start simulating outside of the volume

This commit adds a propagation to the top volume for such particles: For
particles we want to give directly to Geant4 (e.g. they don't have a
mother or their mother isn't simulated) we now first check if the vertex
is inside the top volume and if not if we would hit it in time for this
particle.

If not the particle is skipped, if yes the vertex is generated on the
boundary of the simulation volume (without considering any materials or
fields) and the time shift is applied to the propagation time.

Particles which are added as daughters of other particles given to
Geant4 don't need extra treatment: We know that Geant4 will place them
at the decay vertex of the parent or skip them if we leave the
simulation volume alltogether.

To keep the decay tree constant we now only set the production vertex
and production time from Geant4 if the particle was placed as a daughter
of another particle, otherwise we know the time/vertex and it might be
outside of Geants range.