Reimplement CRY interface
As Torben was crying out that he needs a working cosmic generator I
looked into this mess and decided ... we might wanna redo that.

New CRY interface:

* get rid of the crysetup text file, just pass the string to CRY.
  We might want to add more module parameters for this
* auto locate the cosmic data files
* get rid of the "keepance box" and only have one acceptance shape
* completely redo acceptance check by using VecGeom for the shape and
  intersection handling
* this trivially allows for advanced shapes for the acceptance: sphere, cylinder, box.
* calculate one common time shift for all particles CRY might generate

When CRY generates particles it does so in a n times n meter plane. We
now take these particles and propagate them backwards to the outside of a n x n x
n box. Then we check if they intersect the acceptance shape and if they
do we keep them there.

Now we depend on PR #6891 to propagate these particles to the simulation
volume for Geant4 but in case that is declined we can trivially do it
here too.

* I updated the example script but **not** the `generators/script/generators.py`
* Someone should add more module parameters for the other values passed
  to CRY like which particles to produce and so on.