Use short format for conluence links
Rationale:
The short links are generated based on the static pageid and thus
remain valid when the page is renamed or moved to another location. Since
confluence is not part of the git-repository, changes on confluence can always
happen and we don't want to follow every name change with a new PR.

Downside:
When looking at the documentation source, it is impossible to see which page a
link refers to without clicking on it, because now the title is not in the link
url.

See:
https://confluence.atlassian.com/confkb/the-differences-between-various-url-formats-for-a-confluence-page-278692715.html