Fix random number initalization on HLT (BII-7212)
Usually we initialize random generator in the input process and stream
the random state as part of the event data. But on HLT
the events are send directly to the worker processes without a random
state so we need to initialize on our own in each worker process.

Otherwise all worker processes on each node generate the same sequence
of random numbers. That sequence should still be fully random as there
is a first initialization of the random state on process start. But it
would introduce correlations.

Now directly after we recieved the event and know the event metadata we
initialze the random state using a seed and the event information so
different processes will produce different numbers.