Add authors to online book prerequisites
Checked git blame to check who wrote the contents. Besides Martin, I saw Kilian appear for the files edited in
this commit, but that was only because of the commit 08db6e94c4366158c
("Update basic structure of lessons; add key point boxes") and I thought that it's only restructuring and not
content so didn't add him.