running update: Extend existing open iov if first revision in staging is the same.
If the staging tag contains the same payload as is already present as
open iov in the running tag then don't close the iov and open a new one
with the same revision. Just close the open iov after where the same
payload in staging would have ended

If the staging payload is open then don't update the iov at all.