Fix crash of express reco in passthrough mode
When HLT runs in passthrough mode there's no HLT trigger result
available. So the express reco cannot depend on it.

This commit adds a new option to the TriggerSkim module to allow setting
a return value in case the trigger result is missing. If this is not set
you will get an error/fatal as before. If it is set you will silently
get the result you set and can then use a different path if necessary.