Remove commented out code (it's an anti-pattern, git stash instead)
Commented out code adds cognitive load when reading source code and makes the
user wonder: Why was it commented out? Does it not work? Is the author planning
to add the code in the comments later? Should this be interpreted as an implicit
TODO/FIXME? Or can it be ignored?

At least a comment why a region of code was commented out would be helpful.
Comments are supposed to answer the question why the code was written the way it
is, not as a substitution for version control.
If you want to keep the comments as a TODO, just start them with a "TODO"
statements, many IDE's will recognize that.

Just in case one want to keep these changes to be able to add them later without
having to write the error messages again, one should just use `git stash` or
quickly create a new branch from which one can push later. It's a cheap thing to
do.

See: https://kentcdodds.com/blog/please-dont-commit-commented-out-code