Fix EventOfDoomBuster on HLT
When merging multiple EventOfDoomBusters in the path the after path
condition wasn't taken over correctly (PR #6887).

In addition the EventErrorFlag module was just added to a shared discard
path so we had both error flags set independent of the actual error.

Now, instead of a path to pass around we have a `event_abort` callable
passed to `add_reconstruction` to set the correct event abort procedure
for HLT and other use cases.