Tidy up udst-related functions
Tidying and renaming of udst-related functions. Move them together into
udst.py (analogous to mdst.py). Rename them for consistency with the
equivalent mdst functions, and put the path first. Keep
modularAnalysis.outputUdst for consistency and backwards-compatibility
(there the path argument is kept where it was). Probagate the changes to
skim package, add a few docstring improvements, add to sphinx.