Remove $VO_BELLE2_SW_DIR fallback in findFile()
This should be handled by the tools to explicitly set the data
directories if the ${VO_BELLE2_SW_DIR}/${dataType}-data directory
exists (or just independent on its existence to make sure the variable
is always present).

We could also just remove the fallback and default as this didn't work
correctly before due to a typo and was only fixed in PR #7345 so impact
on users would be minimal but then b2install-data would not
install the data into a directory that would be found automatically.