Fix disabling reading of Branches (BII-7492)
It turns out our efforts to only read the branches requested by the user
didn't work correctly due to some weird naming scheme requirement for
`TTree::SetBranchStatus`.

This was mostly not a problem but it can cause problems if the objects
in the branches on file are not compatible with the current software for
whatever reasons, e.g. in cdst processing.

Luckily we don't want or need a weird regular expression based name
matching as we know the tree structure and can thus do this
programmatically ourselves by just disabling the branch and all its
children branches.