Heavy lifting of the coding style.
The main modification is related to the raw-data StoreArrays. They are moved to private members of the module. In this way, they are created once (by the constructor) and not once per event. Moreover, I replaced the index-based for loops with the range-based for loops.

The other modifications are related to Doxygen, Belle II coding rules and ordering of the headers.