Fix the gcc and intel compilation warnings in the trg package.
When I fixed the clang warnings (by using #pragma clang), I didn't pay attention to the possibility to introduce gcc and intel warnings, and this was the case. #pragma clang is not known by gcc or intel, so let me silence the clang warnings only when clang is actually used (thus, when __clang__ is defined).