2dtracks/cc unpacked 6->4, fix missing initialization
The loop in line 691 in decode2DOutput loops over the range if
nMax2DTracksPerSL and uses the same index to access the posTrack array,
which has only a length of 4.

the foundoldtrack bool array is only filled in newer versions. It was
not initialized and used in a B2DEBUG message, where a std::to_string is
called and throws an error