reactivate the tests of analysis examples and tutorials
The use of boost's canonical seems to remove the trailing slash as well so that the scripts for those tests were no longer found and the test finished successfully (without actually running anything). There might be other instances in the software where this changed behaviour of find_file as an effect.