Define a proper rawFormat for cDSTs to be used with MC generated samples.
If for data the rawFormat is basically raw data + tracking, for MC we can't use the same scheme, because the packers do not keep the MC truth information. Thus, for MC the minimal rawFormat we can define is digits + tracking, and this is what it is added here.

To use the new format, one has to use `add_cdst_output` setting `rawFormat=True` and `mc=True`. Automatically, the digits + tracking objects are stored in the output file.

I also did some cleanup:
* I defined all the pre-defined lists of objects, in order to call them across modules if necessary (especially for mDST objects, now defined in a single place);
* I removed from mDST objects and from the additional `MCParticles` objects the relationships: these are automatically added by `RootOutput` (I checked it), so it's not necessary to add them.

Reguarding the cleanup: should I move the lists of objects in a separate file? It would be cleaner, but I don't have strong opinions.