Fix MDST compatibility tests
While the mapping from bit name to number is now public it doesn't test
any mdst content, just a convenient interface to the DB mapping. As such
I think it is better to not test it here as this would introduce a
strong dependence on the globaltag.

For example, the mapping doesn't exist in the globaltag before release 2
so it cannot be tested there.