Fix setting of production vertex/time for secondaries (BII-8127)
With 5.1 we introduced a feature that particles can start outside the
simulation volume and are propagated to the boundary of the simulation
volume automatically.

In the process we introduced a regression that the production vertex and
time for secondaries are always set to zero.

This commit restores the correct behavior.