Revamp notebook integration
The hep_ipython_tools are too heavy to be used as default and the
widgets don't necessarily work, especially with Jupyterhub servers.

So let's keep it simple and just have a trivial nanny process for the
process() call but give some extra information.

Also, make Module and Path use the html when printing them