Add rich formatting for ProcessStatistics
This commit adds rich formatting to the process statistics object in
Jupyter Notebooks.

For this we slightly restructure how ProcessStatisticsPython works: We
no longer just return strings but new ProcessStatisticsPython objects
when the user selects a different counter/module list. Then jupyter
can use the type information to show the correct output.

Also, this is a thin wrapper around the ProcessStatistics object in the
DataStore and doesn't need to be a singleton.

To make this actually useful in Jupyter notebooks we transfer the
ProcessStatistics object back from the nanny process and put it in the
DataStore. This might look like a bit overkill but is mostly a proof of
concept how this could be extended to further objects if necessary to
transfer state from the execution process to the main process.