Script to convert Pythia codes in .dec files.
Once the externals are updated to v01-10-00 this can be run on all dec
files and the conversion can be disabled in
`generators/evtgen/src/EvtGenInterface.cc:63`.

However I suspect that there will be a lot of people with dec files not
in the repository so this needs to be very carefully coordinated and
announced so that people are actually aware that they need to convert
their custom dec files.

The script adds a comment about the conversion to the top of the dec
file and will not convert the codes again if the comment can be found.
However if the comment is removed/modified the modification could be run
a second time ... which is not correct.