Allow individual skims in b2skim-prod.
In get_skim_stats_dict, rather than raising an error, set CombinedSkimDefinitions to an empty dict if the CombinedSkims does not exist.
Elsewhere, I then check if the skim name exists in the registry, and if so, take that individual skim to be the skim we're producing.
This will work under the assumption that no combined skim name conflicts with an individual skim name.