User new clearer code example provided by Sam Cunliffe
I removed the "WILL NOT WORK" section, because I don't think that it's useful
from a pedagogical point of view to provide examples of what does not work (at
least this is what I learnt in about sports training). Also it adds verbosity.
Better keep the examples short and to the point.