Fix usage of testing payloads (BII-88416)
Changing the creation of payloads to use the checksum as revision
introduced a bug.

Central payloads work fine and testing payloads work fine but if testing
payloads are used together with central payloads the testing payloads is
overwritten by the central payload (as the revision is now zero).

The fix is simple but a bit cryptic as we need to check for revision and
some other property in the MetadataProvider. There might be a nicer
solution like adding a property like `found` to the PayloadMetadata to
check against instead of relying on some of the members.