Added --analysis-globaltag argument to b2skim-stats-submit, and added dry-run before submission.
Because this script only submits bsub jobs (and doesn't do any processing on its own), the user is
only notified that something went wrong with their scripts when they check the log files. I added a
dry-run before submission is done, so that the user is notified immediately if they forgot to
include a analysis global tag where one was required.