Change type mismatch detection
Checking the types via a dummy evaluation with nullptr as particle caused tests to
fail.
Instead we annotate m_functions with the registered type information as a std::pair.
At function evaluation, we compare the result data type with the registered type
before we fill the tree. If there is a mismatch, a warning is generated.