Implement Frank's suggestion regarding special binary electron ID variable names.
Change the API for the standard lepton lists creation to allow passing a release number to set the proper variable names and aliases, and now users should explicitly pass the name of the LID GT to the helper as well.

Add an example script for application of LID weights in basf2.

Add special LID variables and the BDT to the pid collection.