Make documentation clear about direction of CurlTagger pT cut
I wasn't sure whether it's an upper cut or lower cut. I actually had to look
into the C++ source code of the module to make sure that only tracks with a
pT **below** that cut are considered as curler candidates. I think in hindsight
this is obvious, as we don't expect high-momentum tracks to be curlers, but I
would prefer the documentation to be clear about that for people who are not
experts.