Optimises search, handles Gitlab config file check
Search for related issues is optimised with a direction reversal.
Is much faster than the previous version.

A check has been added to verify access to Gitlab config file
where the access token is sured. This will toggle integration
features. This is a temporary fix until a solution for easier
access to the file is determined.