Implement neural network based PID variables
Implement `pidNeuralNetworkValueExpert` variable, which calculates the
probability for a given particle-species hypothesis using a neural
network. The input variables to this network are high-level information,
such as the likelihoods from the subdetectors for PID for all 6
hypotheses and the track momentum and charge.

The neural-network and its parameters are loaded from the conditions
data base. This allows to use different neural networks using the same
basf2 code.

Also, implement `PIDNeuralNetworkParametersCreator` module used to store the neural
network parameters in the conditions  data base.