Use momentum from TrackFitResult insteas of `frame.getMomentum(part)`
If the analyst would perform a vertex fit or momentum scaleing, the
momentum of the particle will be change and one may use inconsistent
momenta for the evaluation of the network.
However, `TrackFitResult::getMomentum()` does not change in this cases.

Additionally, prepared the PIDNeuralNetwork class to calculate the
probabilities for all hypotheses at once. This is not used at the
moment, but may be used in the future to minimize the number of neural
network evaluations.