In ParticleLoaderModule.cc there is a problem for loading neutral clusters of particle which are not self conjugated. There is no cluster information based on which a cluster can be assigned to a particle or an anti-particle. However what the particle loader did to this point is it registered both the particle and the anti-particle list and filled only one!
This can be easily seen by trying to ma.fillParticleList('n0:all') and/or ma.fillParticleList('anti-n0:all'),
both particle lists will be registered, but depending on which one will be explicitly written one will be filled and the other empty.
Calling the second ma.fillParticleList on the anti-particle list in succession, also does not help, as the anti-particle list is already registered and will therefore not be filled again.

Simplest fix (I suppose) to not automatically register the anti-particle list for neutral clusters, so If one wants to load both the particle list and anti-particle list,
they need to load them separately and combine them later if they wish?